requirements_dist.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
future==0.18.2
idna==3.4
iso8601==1.1.0
Pillow==9.3.0
pyserial==3.5
PyYAML==6.0
requests==2.28.1
serial==0.0.97
soupsieve==2.3.2.post1
urllib3==1.26.13
